+++ app/views/well_known/host_meta/show.xml.ruby
