+++ app/views/well_known/webfinger/show.json.rabl
