+++ app/views/well_known/webfinger/show.xml.ruby
